Syncable class moved outside of adapter
